Cleaned up Home.js by separating frontend and backend logic (new folder)
